+++ CHANGELOG.md
@@ -8,6 +8,7 @@ bugs in actually-released versions.
 
 - Add MQTT support. Flamenco Manager can now send internal events to an MQTT broker.
 - Simplify the preview video filename when a complex set of frames rendered ([#104285](https://projects.blender.org/studio/flamenco/issues/104285)). Instead of `video-1, 4, 10.mp4` it is now simply `video-1-10.mp4`.
+- Make the `blendfile` parameter of a `blender-render` command optional. This makes it possible to pass, for example, a Python file that loads/constructs the blend file, instead of loading one straight from disk.
 
 ## 3.4 - released 2024-01-12